Rename the file abstract motion controller
