+++ src/XR/motionController/index.ts
@@ -1,4 +1,4 @@
-export * from "./webXRAbstractController";
+export * from "./webXRAbstractMotionController";
 export * from "./webXRControllerComponent";
 export * from "./webXRGenericMotionController";
 export * from "./webXRMicrosoftMixedRealityController";
