+++ src/XR/webXRInputSource.ts
@@ -3,7 +3,7 @@ import { AbstractMesh } from "../Meshes/abstractMesh";
 import { Quaternion, Vector3 } from '../Maths/math.vector';
 import { Ray } from '../Culling/ray';
 import { Scene } from '../scene';
-import { WebXRAbstractMotionController } from './motionController/webXRAbstractController';
+import { WebXRAbstractMotionController } from './motionController/webXRAbstractMotionController';
 import { WebXRMotionControllerManager } from './motionController/webXRMotionControllerManager';
 
 let idCount = 0;
